Add py.test and coverage + coveralls build reports.
