+++ .gitignore
@@ -1,1 +1,3 @@
 *.pyc
+.coverage
+venv/
