+++ .travis.yml
@@ -1,9 +1,13 @@
 language:
     - python
 python:
-    - "2.6"
-    - "2.7"
+    - '2.6'
+    - '2.7'
 install:
-    - "pip install -r requirements.txt"
+    - pip install -r requirements.txt
+    - pip install -r requirements-dev.txt
+    - pip install coveralls
 script:
-    - "python src/unittests.py"
+    - py.test --cov-report term-missing --cov=src src
+after_success:
+  coveralls
